requirements.txt
clip==1.0
config==0.5.1
einops==0.8.1
jsonlines==3.1.0
librosa==0.10.1
lmdb==1.4.1
matplotlib==3.5.3
mmcv==2.1.0
modules==1.0.0
numpy==1.22.0
opencv_python==4.8.0.76
Pillow==11.1.0
pyarrow==1.0.1
pytorch_lightning==2.0.0
scikit_learn==1.0.2
six==1.16.0
soundfile==0.12.1
timm==0.3.2
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.66.4
transformers==4.34.0
